requirements.txt
flask
python-dotenv
psycopg2-binary
flask-jwt-extended
gunicorn
flask-restful